+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -779,9 +779,7 @@ export class SnapKeyring {
     snapId: SnapId,
   ): InternalAccount['metadata']['snap'] | undefined {
     const snap = this.#getSnap(snapId);
-    return snap
-      ? { id: snapId, name: snap.manifest.proposedName, enabled: snap.enabled }
-      : undefined;
+    return snap ? { id: snapId } : undefined;
   }
 
   #transformToInternalAccount(